fix: use type int for CampaignId

+++ src/Helldivers-2-Models/ArrowHead/Status/PlanetEvent.cs
@@ -24,6 +24,6 @@ public sealed record PlanetEvent(
     long MaxHealth,
     long StartTime,
     long ExpireTime,
-    long CampaignId,
+    int CampaignId,
     List<int> JointOperationIds
 );
